Move types' and interfaces' declarations to lib
